+++ mcpgateway/services/grpc_service.py
@@ -15,8 +15,6 @@ retrieval, updates, activation toggling, and deletion.
 import asyncio
 import base64
 from datetime import datetime, timezone
-import ipaddress
-from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 try:
@@ -38,6 +36,7 @@ from sqlalchemy import and_, delete, desc, select, update
 from sqlalchemy.orm import Session
 
 # First-Party
+from mcpgateway import translate_grpc
 from mcpgateway.config import settings
 from mcpgateway.db import EmailTeam
 from mcpgateway.db import GrpcService as DbGrpcService
@@ -49,14 +48,13 @@ from mcpgateway.services.logging_service import LoggingService
 from mcpgateway.services.team_management_service import TeamManagementService
 from mcpgateway.utils.create_slug import slugify
 from mcpgateway.utils.display_name import generate_display_name
+from mcpgateway.utils.grpc_validation import GrpcServiceError, _validate_grpc_target, _validate_tls_path
 from mcpgateway.utils.pagination import unified_paginate
 
 # Initialize logging
 logging_service = LoggingService()
 logger = logging_service.get_logger(__name__)
 
-
-_GRPC_DISALLOWED_SCHEMES = ("unix:", "unix-abstract:", "vsock:", "fd:")
 
 # DoS guards for descriptors returned by ``grpc.reflection`` — intentionally hardcoded
 # (not exposed via settings) so a config change cannot silently weaken these limits.
@@ -109,106 +107,6 @@ def _validate_reflected_tool_name(tool_name: str) -> None:
         SecurityValidator.validate_tool_name(tool_name)
     except ValueError as exc:
         raise GrpcServiceError(f"Reflected tool name '{tool_name}' rejected: {exc}") from exc
-
-
-def _validate_grpc_target(target: str) -> None:
-    """Validate a gRPC target address against SSRF-unsafe destinations.
-
-    Consults the platform SSRF settings (``ssrf_allow_localhost``,
-    ``ssrf_allow_private_networks``, ``ssrf_allowed_networks``,
-    ``ssrf_blocked_networks``, ``ssrf_blocked_hosts``) so that gRPC
-    targets follow the same rules as HTTP URLs validated by
-    ``SecurityValidator.validate_url``.
-
-    Args:
-        target: gRPC target string. Accepts ``host:port``, bracketed
-            ``[ipv6]:port``, and gRPC name-resolver forms ``dns:///host:port``,
-            ``ipv4:host:port``, ``ipv6:host:port``. Local-only schemes
-            (``unix:``, ``unix-abstract:``, ``vsock:``, ``fd:``) are always
-            rejected because they bypass the network-level SSRF model.
-
-    Raises:
-        GrpcServiceError: If the target uses a forbidden scheme or resolves to a blocked address.
-    """
-    if not target:
-        raise GrpcServiceError("Empty gRPC target address")
-
-    # Local-only schemes bypass the IP-based SSRF model entirely; reject outright.
-    lowered = target.lower()
-    for scheme in _GRPC_DISALLOWED_SCHEMES:
-        if lowered.startswith(scheme):
-            raise GrpcServiceError(f"gRPC target scheme '{scheme.rstrip(':')}' is not permitted")
-
-    # Strip recognised gRPC name-resolver scheme prefixes so the host check below sees a bare host:port.
-    for scheme_prefix in ("dns:///", "dns://", "dns:", "ipv4:", "ipv6:"):
-        if lowered.startswith(scheme_prefix):
-            target = target[len(scheme_prefix) :]
-            break
-
-    # Extract host (strip port). Bracketed IPv6 literals: ``[::1]:50051``.
-    if target.startswith("["):
-        end = target.find("]")
-        if end < 0:
-            raise GrpcServiceError(f"Malformed bracketed gRPC target: {target!r}")
-        host = target[1:end]
-    else:
-        host = target.rsplit(":", 1)[0].strip("[]")
-    if not host:
-        raise GrpcServiceError("Empty gRPC target address")
-
-    # Reserved / multicast IP literals are unconditionally blocked. SecurityValidator._validate_ssrf
-    # only checks blocked-networks / localhost / private; it does not flag is_reserved/is_multicast,
-    # so this guard runs before delegation to keep the original gRPC-validator semantics. Loopback is
-    # excluded because Python flags ``::1`` as both is_loopback AND is_reserved; loopback policy is
-    # handled by SecurityValidator below via ssrf_allow_localhost.
-    try:
-        addr = ipaddress.ip_address(host)
-    except ValueError:
-        addr = None
-    if addr is not None and not addr.is_loopback and (addr.is_reserved or addr.is_multicast):
-        raise GrpcServiceError(f"gRPC target address '{host}' is blocked (reserved/multicast)")
-
-    # Delegate the hostname/IP-network/DNS-resolution policy to the shared SecurityValidator
-    # so gRPC and HTTP follow the same SSRF rules and a hostname like ``metadata.google.internal``
-    # is resolved before being allowed through.
-    # First-Party
-    from mcpgateway.common.validators import SecurityValidator  # pylint: disable=import-outside-toplevel
-
-    if getattr(settings, "ssrf_protection_enabled", True):
-        try:
-            SecurityValidator._validate_ssrf(host, "gRPC target")  # pylint: disable=protected-access
-        except ValueError as exc:
-            raise GrpcServiceError(str(exc)) from exc
-
-
-def _validate_tls_path(path_str: str, label: str = "TLS path") -> Path:
-    """Validate that a TLS cert/key path is within allowed directories.
-
-    Args:
-        path_str: The file path to validate.
-        label: Label for error messages.
-
-    Returns:
-        Resolved Path object.
-
-    Raises:
-        GrpcServiceError: If the path escapes allowed directories.
-    """
-    resolved = Path(path_str).resolve()
-    # Allow only paths under /certs/, /etc/ssl/, /etc/pki/, or the CWD/certs dir
-    allowed_prefixes = (
-        Path("/certs").resolve(),
-        Path("/etc/ssl").resolve(),
-        Path("/etc/pki").resolve(),
-        Path.cwd().joinpath("certs").resolve(),
-    )
-    if not any(resolved.is_relative_to(prefix) for prefix in allowed_prefixes):
-        raise GrpcServiceError(f"{label} '{path_str}' is outside allowed certificate directories")
-    return resolved
-
-
-class GrpcServiceError(Exception):
-    """Base class for gRPC service-related errors."""
 
 
 class GrpcServiceNotFoundError(GrpcServiceError):
@@ -960,10 +858,6 @@ class GrpcService:
         if not service.enabled:
             raise GrpcServiceError(f"Service '{service.name}' is disabled")
 
-        # Import here to avoid circular dependency
-        # First-Party
-        from mcpgateway.translate_grpc import GrpcEndpoint  # pylint: disable=import-outside-toplevel,cyclic-import
-
         # Parse method name (service.Method format)
         if "." not in method_name:
             raise GrpcServiceError(f"Invalid method name '{method_name}', expected 'service.Method' format")
@@ -986,7 +880,7 @@ class GrpcService:
         stored_descriptors = discovered.get("_file_descriptors", [])
         has_stored_descriptors = bool(stored_descriptors)
 
-        endpoint = GrpcEndpoint(
+        endpoint = translate_grpc.GrpcEndpoint(
             target=service.target,
             reflection_enabled=not has_stored_descriptors,
             tls_enabled=service.tls_enabled,
