+++ tests/unit/mcpgateway/services/test_grpc_service_no_grpc.py
@@ -960,7 +960,7 @@ async def test_perform_reflection_tls_reads_cert_and_key(monkeypatch, service, d
         visibility="public",
     )
 
-    with patch("mcpgateway.services.grpc_service.Path.read_bytes", side_effect=[b"cert", b"key"]):
+    with patch("mcpgateway.utils.grpc_validation.Path.read_bytes", side_effect=[b"cert", b"key"]):
         await service._perform_reflection(db, db_service)
 
     assert db_service.reachable is True
@@ -997,7 +997,7 @@ async def test_perform_reflection_tls_missing_cert(monkeypatch, service, db):
         visibility="public",
     )
 
-    with patch("mcpgateway.services.grpc_service.Path.read_bytes", side_effect=FileNotFoundError("missing")):
+    with patch("mcpgateway.utils.grpc_validation.Path.read_bytes", side_effect=FileNotFoundError("missing")):
         with pytest.raises(GrpcServiceError):
             await service._perform_reflection(db, db_service)
